requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
SQLAlchemy>=2.0.23
Werkzeug>=3.0.1
python-dotenv>=1.0.0
flask-login>=0.6.3
WTForms>=3.1.1
Flask-WTF>=1.2.1
openpyxl>=3.1.2
xlsxwriter>=3.1.9
pandas>=2.1.3
numpy>=1.26.2
xlrd>=2.0.1
pyotp>=2.9.0
psycopg2-binary>=2.9.9
Flask-Caching>=2.1.0
Flask-Migrate>=4.0.5
reportlab>=4.1.0
gunicorn>=21.2.0
gevent>=24.2.1 